Implement multi-file upload and better download action
Previously, the user was only able to send one file at a time. Now the user can select a range of file to send to the server. Additionally, I improved the downlaod button performance wise and complexity wise. Another great possibility is to use 'content-disposition: attachment', which I will probably implement in a future iteration. In an off note, the file structure of the project is a bit messy now, so I need a way to separate *.js and *.ts perhaps in src/ and dist/ folders, and use a better auto_compiler code base which will make the typescript automatic compilation process more robust.